Reduced duplicate code on ui implementation
