+++ src/platform/guinone.cpp
@@ -5,7 +5,6 @@
 //-----------------------------------------------------------------------------
 #include "solvespace.h"
 #include <cairo.h>
-#include <dirent.h>
 
 namespace SolveSpace {
 
@@ -127,22 +126,7 @@ std::vector<Platform::Path> GetFontFiles() {
     // If a custom fonts folder was specified via --fonts-folder, scan it
     // instead of the system fonts. This is per-session and not persisted.
     if(!customFontsFolder.empty()) {
-        std::vector<Platform::Path> fonts;
-        DIR *dir = opendir(customFontsFolder.c_str());
-        if(dir != NULL) {
-            struct dirent *entry;
-            while((entry = readdir(dir)) != NULL) {
-                std::string name = entry->d_name;
-                if(name == "." || name == "..") continue;
-                Platform::Path fontPath =
-                    Platform::Path::From(customFontsFolder).Join(name);
-                if(fontPath.HasExtension("ttf") || fontPath.HasExtension("otf")) {
-                    fonts.push_back(fontPath);
-                }
-            }
-            closedir(dir);
-        }
-        return fonts;
+        return Platform::GetFontsFromDirectory(customFontsFolder);
     }
     return fontFiles;
 }
